Extend burst-read to CC2510 and add fallback
The CC2510 chip can benefit from the same performance boost as the
CC254x chips, as they share the same instruction set. Change to use the
burstReadCode function and also add a fallback so if the probe device
isn't running with support for the 0xF3 BURST_RD instruction we can
fallback to the old slower method.